Finished initial implementation of AVLTreeDictionary.
At first, I though that the ICollection<T>s that were used for the Keys
and Values would need more work due to their constant connection to
their dictionary, but it turns out that wasn't needed thanks to
dictionary containing the IEnumerable<KeyValuePair<TKey, TValue>>
interface inside of the IDictionary<TKey, TValue> interface. Having that
implemented into the dictionary made it so that I didn't have to expose
the internals of the dictionary to the collection classes.